+++ R/gg_grouped_y2.R
@@ -105,15 +105,17 @@ if(
     chart_height < 5.5 & direction == 'vertical' ~  (max_y_val + (max_y_val / (chart_height * 2)) * 2),
     TRUE ~  (max_y_val + max_y_val / 10) #direction == 'vertical'
   )
-  nudge_y <- dplyr::case_when(
-    direction == 'horizontal' ~ 0.5, # places the percent_label in the middle of the bar
+  nudge <- dplyr::case_when(
     nudge != 0 ~ nudge,
+    TRUE ~ max_y_val/100
+  )
+  nudge_y <- dplyr::case_when(
+    direction == 'horizontal' ~ 0.5,
     direction == 'vertical' ~ (max_y_val/(max_y_val * 5)) * -1
   )
   nudge_x <- dplyr::case_when(
-    direction == 'vertical' ~ 0.5,
-    nudge != 0 ~ nudge,
-    direction == 'horizontal' ~ (max_y_val / (max_y_val * 4) + str_add) * -1
+    direction == 'vertical' ~ 0.5, # places the percent_label in the middle of the bar for
+    direction == 'horizontal' ~ 0 # Left justifies for horizontal
   )
   label_length <- dplyr::case_when(
     label_length != 45 ~ label_length,
@@ -154,12 +156,12 @@ if(
     ggplot2::geom_text(
       ggplot2::aes(
         label = !!label_flag,
-        color = !!color_flag
+        color = !!color_flag,
+        y = !!y_flag + nudge
       ),
       family = font_family,
       size = label_size,
-      # nudge_y = nudge, # grr, doesn't work with position argument. Have to do v/hjust instead
-      position = ggplot2::position_dodge(width = 0.9),
+      position = ggplot2::position_dodge2(width = 0.9),
       hjust = nudge_x,
       vjust = nudge_y
     ) +